exercises(sum-of-multiples): use u64 for sum
A u64 is sufficient for the maximum value of the sum when the limit is a
u32. The maximum sum occurs for

    const factors = [_]u32{1};
    const limit = 4_294_967_295 // the maximum value of a u32

and resulting sum of the integers from 1 to 4_294_967_295 is:

     9_223_372_034_707_292_160

which is less than the maximum value of a u64:

    18_446_744_073_709_551_615

+++ exercises/practice/sum-of-multiples/test_sum_of_multiples.zig
@@ -4,7 +4,7 @@ const testing = std.testing;
 const sum = @import("sum_of_multiples.zig").sum;
 
 test "no multiples within limit" {
-    const expected: u32 = 0;
+    const expected: u64 = 0;
     const factors = [_]u32{ 3, 5 };
     const limit = 1;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -12,7 +12,7 @@ test "no multiples within limit" {
 }
 
 test "one factor has multiples within limit" {
-    const expected: u32 = 3;
+    const expected: u64 = 3;
     const factors = [_]u32{ 3, 5 };
     const limit = 4;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -20,7 +20,7 @@ test "one factor has multiples within limit" {
 }
 
 test "more than one multiple within limit" {
-    const expected: u32 = 9;
+    const expected: u64 = 9;
     const factors = [_]u32{3};
     const limit = 7;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -28,7 +28,7 @@ test "more than one multiple within limit" {
 }
 
 test "more than one factor with multiples within limit" {
-    const expected: u32 = 23;
+    const expected: u64 = 23;
     const factors = [_]u32{ 3, 5 };
     const limit = 10;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -36,7 +36,7 @@ test "more than one factor with multiples within limit" {
 }
 
 test "each multiple is only counted once" {
-    const expected: u32 = 2318;
+    const expected: u64 = 2318;
     const factors = [_]u32{ 3, 5 };
     const limit = 100;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -44,7 +44,7 @@ test "each multiple is only counted once" {
 }
 
 test "a much larger limit" {
-    const expected: u32 = 233_168;
+    const expected: u64 = 233_168;
     const factors = [_]u32{ 3, 5 };
     const limit = 1000;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -52,7 +52,7 @@ test "a much larger limit" {
 }
 
 test "three factors" {
-    const expected: u32 = 51;
+    const expected: u64 = 51;
     const factors = [_]u32{ 7, 13, 17 };
     const limit = 20;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -60,7 +60,7 @@ test "three factors" {
 }
 
 test "factors not relatively prime" {
-    const expected: u32 = 30;
+    const expected: u64 = 30;
     const factors = [_]u32{ 4, 6 };
     const limit = 15;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -68,7 +68,7 @@ test "factors not relatively prime" {
 }
 
 test "some pairs of factors relatively prime and some not" {
-    const expected: u32 = 4419;
+    const expected: u64 = 4419;
     const factors = [_]u32{ 5, 6, 8 };
     const limit = 150;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -76,7 +76,7 @@ test "some pairs of factors relatively prime and some not" {
 }
 
 test "one factor is a multiple of another" {
-    const expected: u32 = 275;
+    const expected: u64 = 275;
     const factors = [_]u32{ 5, 25 };
     const limit = 51;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -84,7 +84,7 @@ test "one factor is a multiple of another" {
 }
 
 test "much larger factors" {
-    const expected: u32 = 2_203_160;
+    const expected: u64 = 2_203_160;
     const factors = [_]u32{ 43, 47 };
     const limit = 10_000;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -92,7 +92,7 @@ test "much larger factors" {
 }
 
 test "all numbers are multiples of 1" {
-    const expected: u32 = 4_950;
+    const expected: u64 = 4_950;
     const factors = [_]u32{1};
     const limit = 100;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -100,7 +100,7 @@ test "all numbers are multiples of 1" {
 }
 
 test "no factors means an empty sum" {
-    const expected: u32 = 0;
+    const expected: u64 = 0;
     const factors = [_]u32{};
     const limit = 10_000;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -108,7 +108,7 @@ test "no factors means an empty sum" {
 }
 
 test "the only multiple of 0 is 0" {
-    const expected: u32 = 0;
+    const expected: u64 = 0;
     const factors = [_]u32{0};
     const limit = 1;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -116,7 +116,7 @@ test "the only multiple of 0 is 0" {
 }
 
 test "the factor 0 does not affect the sum of multiples of other factors" {
-    const expected: u32 = 3;
+    const expected: u64 = 3;
     const factors = [_]u32{ 3, 0 };
     const limit = 4;
     const actual = try sum(testing.allocator, &factors, limit);
@@ -124,7 +124,7 @@ test "the factor 0 does not affect the sum of multiples of other factors" {
 }
 
 test "solutions using include-exclude must extend to cardinality greater than 3" {
-    const expected: u32 = 39_614_537;
+    const expected: u64 = 39_614_537;
     const factors = [_]u32{ 2, 3, 5, 7, 11 };
     const limit = 10_000;
     const actual = try sum(testing.allocator, &factors, limit);
